Frontend: refactor into components
